xpUpdate

+++ exp/src/main/java/com/bee/exp/web/dto/TaskResponse.java
@@ -4,16 +4,23 @@ import com.bee.exp.domain.TaskDifficulty;
 import com.bee.exp.domain.TaskStatus;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TaskResponse {
+
     private Long id;
     private String title;
     private String description;
+
     private TaskDifficulty difficulty;
-    private Integer price;
+    private BigDecimal price;
     private TaskStatus status;
+
     private String companyName;
     private String assignedEngineerName;
-    private Boolean claimedByMe;    // Bu junior bu görevi aldı mı?
-    private Boolean submittedByMe;  // Bu junior bu görevi submit etti mi?
+
+    // 👇 Junior UI’nin beklediği alanlar
+    private Boolean claimedByMe;
+    private Boolean submittedByMe;
 }
